models._utils, ROM test framework
